+++ metadata_api/settings.py
@@ -28,9 +28,15 @@ class Settings(BaseSettings):
         "analyses": "analysis.json",
         "utilities": "utility.json",
     }
-    # Resource texts file and collections file (relative to "static" directory)
+    # Other files relative to "static" directory
     RESOURCE_TEXTS_FILE: str = "resource-texts.json"
     COLLECTIONS_FILE: str = "collection.json"
+    LICENSE_INFO_FILE: str = "license-info.json"
+
+    # URL to SPDX licenses JSON data
+    LICENSE_INFO_URL: str = "https://raw.githubusercontent.com/spdx/license-list-data/refs/heads/main/json/licenses.json"
+    # Name used for licenses that are not listed in the above license info
+    LICENSE_OTHER_NAME: str = "LicenseRef-Other"
 
     # Absolute path to directory containing the metadata yaml files (https://github.com/spraakbanken/metadata)
     METADATA_DIR: Path = Path("/home/fksbwww/metadata-api/dev/metadata")